refactor: Rename and relocate roboflow_hf module to improve structure
